git-svn-id: http://svn.openmrs.org/openmrs/tags/latest@27789 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/ObsDAO.java
@@ -82,9 +82,9 @@ public interface ObsDAO {
 	 *      java.lang.Integer, java.lang.Integer, java.util.Date, java.util.Date, boolean)
 	 */
 	public List<Obs> getObservations(List<Person> whom, List<Encounter> encounters, List<Concept> questions,
-	                                 List<Concept> answers, List<PERSON_TYPE> personTypes, List<Location> locations,
-	                                 List<String> sort, Integer mostRecentN, Integer obsGroupId, Date fromDate, Date toDate,
-	                                 boolean includeVoidedObs) throws DAOException;
+	        List<Concept> answers, List<PERSON_TYPE> personTypes, List<Location> locations, List<String> sort,
+	        Integer mostRecentN, Integer obsGroupId, Date fromDate, Date toDate, boolean includeVoidedObs)
+	        throws DAOException;
 	
 	/**
 	 * @see org.openmrs.api.ObsService#getObservationCount(java.util.List, java.util.List,
@@ -92,11 +92,10 @@ public interface ObsDAO {
 	 *      java.util.Date, java.util.Date, boolean)
 	 * @see ObsService#getObservationCount(org.openmrs.ConceptName, boolean)
 	 */
-	public Integer getObservationCount(List<Person> whom, List<Encounter> encounters, List<Concept> questions,
-	                                   List<Concept> answers, List<PERSON_TYPE> personTypes, List<Location> locations,
-	                                   Integer obsGroupId, Date fromDate, Date toDate,
-	                                   List<ConceptName> valueCodedNameAnswers, boolean includeVoidedObs)
-	                                                                                                     throws DAOException;
+	public Long getObservationCount(List<Person> whom, List<Encounter> encounters, List<Concept> questions,
+	        List<Concept> answers, List<PERSON_TYPE> personTypes, List<Location> locations, Integer obsGroupId,
+	        Date fromDate, Date toDate, List<ConceptName> valueCodedNameAnswers, boolean includeVoidedObs)
+	        throws DAOException;
 	
 	/**
 	 * Auto generated method comment
